+++ modules/canvaskit/npm_build/types/index.d.ts
@@ -460,6 +460,7 @@ export interface CanvasKit {
 
     // Factories, i.e. things made with CanvasKit.Foo.MakeTurboEncabulator()
     readonly ParagraphBuilder: ParagraphBuilderFactory;
+    readonly Blender: BlenderFactory;
     readonly ColorFilter: ColorFilterFactory;
     readonly FontCollection: FontCollectionFactory;
     readonly FontMgr: FontMgrFactory;
@@ -1201,6 +1202,11 @@ export interface AnimatedImage extends EmbindObject<"AnimatedImage"> {
      */
     width(): number;
 }
+
+/**
+ * See SkBlender.h for more on this class. The objects are opaque.
+ */
+export type Blender = EmbindObject<"Blender">;
 
 /**
  * See SkCanvas.h for more information on this class.
@@ -2110,6 +2116,18 @@ export interface Paint extends EmbindObject<"Paint"> {
     setBlendMode(mode: BlendMode): void;
 
     /**
+     * Sets the current blender, increasing its refcnt, and if a blender is already
+     * present, decreasing that object's refcnt.
+     *
+     * * A nullptr blender signifies the default SrcOver behavior.
+     *
+     * * For convenience, you can call setBlendMode() if the blend effect can be expressed
+     * as one of those values.
+     * @param blender
+     */
+    setBlender(blender: Blender): void;
+
+    /**
      * Sets alpha and RGB used when stroking and filling. The color is four floating
      * point values, unpremultiplied. The color values are interpreted as being in
      * the provided colorSpace.
@@ -2734,6 +2752,12 @@ export interface RuntimeEffect extends EmbindObject<"RuntimeEffect"> {
     /**
      * Returns a shader executed using the given uniform data.
      * @param uniforms
+     */
+    makeBlender(uniforms: Float32Array | number[] | MallocObj): Blender;
+
+    /**
+     * Returns a shader executed using the given uniform data.
+     * @param uniforms
      * @param localMatrix
      */
     makeShader(uniforms: Float32Array | number[] | MallocObj,
@@ -3344,6 +3368,17 @@ export interface Matrix4x4Helpers {
     transpose(matrix: Matrix4x4 | number[]): number[];
 }
 
+    /**
+     * For more information, see SkBlender.h.
+     */
+export interface BlenderFactory {
+    /**
+     * Create a blender that implements the specified BlendMode.
+     * @param mode
+     */
+    Mode(mode: BlendMode): Blender;
+}
+
 export interface ParagraphBuilderFactory {
     /**
      * Creates a ParagraphBuilder using the fonts available from the given font manager.
@@ -3797,6 +3832,14 @@ export interface RuntimeEffectFactory {
      *                   be printed to console.log().
      */
     Make(sksl: string, callback?: (err: string) => void): RuntimeEffect | null;
+
+    /**
+     * Compiles a RuntimeEffect from the given blender code.
+     * @param sksl - Source code for a blender written in SkSL
+     * @param callback - will be called with any compilation error. If not provided, errors will
+     *                   be printed to console.log().
+     */
+    MakeForBlender(sksl: string, callback?: (err: string) => void): RuntimeEffect | null;
 
     /**
      * Adds debug tracing to an existing RuntimeEffect.