+++ modules/canvaskit/CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   used to store this picture. This size does not include large objects like
   images.
  - `FontMgr.matchFamilyStyle` finds the closest matching typeface to the specified familyName and style.
+- `Paint.setBlender` Sets the current blender.
+- `Blender.Mode` Create a blender that implements the specified BlendMode.
+- `RuntimeEffect.MakeForBlender` Compiles a RuntimeEffect from the given blender code.
 
 ### Fixed
  - `EmbindObject` has been updated to allow TypeScript to differentiate between opaque